requirements.txt
django <= 4.2
djangorestframework
django-cors-headers == 3.14.0
djangorestframework-simplejwt
requests
psycopg2-binary
dj-database-url == 1.2.0 
django-dotenv